requirements.txt
mouse~=0.7.1
pyautogui~=0.9.53
Pillow~=9.4.0